[AN] refactor: SpeedAdapter를 별도 파일로 분리 #688
